Added tentative Location
This PR is to record some discussion on Location.

How do we want to treat Location?  I'm thinking we just reference an existing vocabulary's term, but haven't researched what we can use.  

This model shows that Location is different than Resource.

Location would need to be definable by the users to the level of detail they would need.  For example: SW corner of elf's basement.

Discussion?  Disagreement?