New resource UML pic
For discussion, not looking for merge necessarily, although it probably reflects where we are at better than the existing one. This is mostly an attempt to document what we have thus far.   I'd like a reading on what are our remaining issues on vf:Resource, and/or issues with this model.  And I do expect there is still some discussion on what I have here. :)